consolidated models and got rid of redundant fields
